+++ taskw/test/storage_test.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:test/test.dart';
@@ -8,7 +9,23 @@ import 'package:taskc/taskc.dart';
 import 'package:taskw/taskw.dart';
 
 void main() {
-  test('test profiles', () {
+  var credentialsKey = Credentials.fromString(
+    parseTaskrc(
+      File(
+        '../fixture/.taskrc',
+      ).readAsStringSync(),
+    )['taskd.credentials'],
+  ).key;
+
+  var txData = File(
+    '../fixture/var/taskd/orgs/Public/users/$credentialsKey/tx.data',
+  );
+
+  test('test profiles', () async {
+    if (txData.existsSync()) {
+      txData.deleteSync();
+    }
+
     var storage = Storage(
       Directory(
         'test/profile-testing/storage/${const Uuid().v1()}',
@@ -16,31 +33,33 @@ void main() {
     );
     [
       Task(
-        uuid: 'foo',
+        uuid: const Uuid().v1(),
         status: 'pending',
-        description: 'test',
+        description: 'foo',
         entry: DateTime.now(),
-        tags: const ['baz'],
+        modified: DateTime.now(),
+        tags: const ['qux'],
       ),
       Task(
-        uuid: 'bar',
+        uuid: const Uuid().v1(),
         status: 'waiting',
-        description: 'test',
+        description: 'bar',
         entry: DateTime.now(),
+        modified: DateTime.now(),
         wait: DateTime.now(),
       ),
       Task(
-        uuid: 'baz',
+        uuid: const Uuid().v1(),
         status: 'pending',
-        description: 'test',
+        description: 'baz',
         entry: DateTime.now(),
+        modified: DateTime.now(),
         until: DateTime.now(),
       ),
     ].forEach(storage.mergeTask);
     storage
       ..tags()
-      ..allData()
-      ..getTask('foo');
+      ..allData();
     for (var entry in {
       '.taskrc': '.taskrc',
       'taskd.ca': '.task/ca.cert.pem',
@@ -54,6 +73,39 @@ void main() {
         contents: File('../fixture/${entry.value}').readAsStringSync(),
       );
     }
-    storage.synchronize();
+
+    try {
+      await storage.synchronize();
+    } on BadCertificateException catch (e) {
+      storage.addFileContents(
+        key: 'server.cert',
+        contents: e.certificate.pem,
+      );
+    }
+
+    await storage.synchronize();
+
+    for (var data in ['backlog', 'pending', 'completed']) {
+      var dataFile = File('../fixture/.task/$data.data');
+      if (dataFile.existsSync()) {
+        dataFile.deleteSync();
+      }
+    }
+
+    Process.runSync('task', ['sync'],
+        workingDirectory: '..', environment: {'HOME': 'fixture'});
+    Process.runSync('task', [],
+        workingDirectory: '..', environment: {'HOME': 'fixture'});
+    var cliExport = Process.runSync('task', ['export'],
+        workingDirectory: '..', environment: {'HOME': 'fixture'}).stdout;
+    Process.runSync('task', ['sync'],
+        workingDirectory: '..', environment: {'HOME': 'fixture'});
+
+    await storage.synchronize();
+
+    var libExport = storage.export();
+
+    expect(json.decode(libExport), json.decode(cliExport));
+    expect(libExport, cliExport);
   });
 }